\name{FreqProfR-package}
\alias{FreqProfR-package}
\alias{FreqProfR}
\docType{package}
\title{
Frequency Profiles computing and plotting.
}
\description{
This package aims at computing and plotting frequency profiles from data encoded
into .bin, .fpw and .csv formats.

It has a user-friendly interface - using the "Shiny App" framework - so that
anyone can use it, even without any prior on R programming.
}
\details{
\tabular{ll}{
Package: \tab FreqProfR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-11-04\cr
}

}
\author{
Thomas Boulier <thomas.boulier@polytechnique.edu>

Maintainer: Ronald Robertson <rrobertson@aibrt.org>
}
\examples{
plot.freqprof(freqprof(import.data()))
}
