% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FreqProf_sourcecode.R
\name{head.freqprof}
\alias{head.freqprof}
\title{Prints the head of the data contained into a variable of class freqprof}
\usage{
\method{head}{freqprof}(data.freqprof)
}
\arguments{
\item{data.freqprof}{the data}
}
\value{
NULL.
}
\description{
Prints the head of the data contained into a variable of class freqprof
}
\examples{
head.freqprof(freqprof(import.data()))
}

