% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FreqProf_sourcecode.R
\name{read.bin}
\alias{read.bin}
\title{Reads the data in the file "filename", which is supposed to be a .bin file}
\usage{
read.bin(filename)
}
\arguments{
\item{filename}{a string indicating the path of the file.}
}
\value{
A data.frame giving the raw data.
}
\description{
Reads the data in the file "filename", which is supposed to be a .bin file
}
\examples{
read.bin(file.choose())
}

